sideAttack() created in fight.c
